ehrbase/project_management#143
Fixes EHR endpoint handling of prefer header and the matching responses.
