+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/response/EhrResponseData.java
@@ -22,6 +22,10 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
 import com.nedap.archie.rm.ehr.EhrStatus;
 import com.nedap.archie.rm.support.identification.HierObjectId;
+import org.ehrbase.api.dto.CompositionDto;
+import org.ehrbase.api.dto.ContributionDto;
+
+import java.util.List;
 
 /**
  * Basic set of response data regarding EHR operations. Used as default or when `PREFER` header requests minimal response.
@@ -38,6 +42,10 @@ public class EhrResponseData {
     private EhrStatus ehrStatus;
     @JsonProperty(value = "time_created")
     private String timeCreated;
+    @JsonProperty
+    private List<CompositionDto> compositions;
+    @JsonProperty
+    private List<ContributionDto> contributions;
 
     public HierObjectId getSystemId() {
         return systemId;
@@ -71,5 +79,20 @@ public class EhrResponseData {
         this.timeCreated = timeCreated;
     }
 
+    public List<CompositionDto> getCompositions() {
+        return compositions;
+    }
+
+    public void setCompositions(List<CompositionDto> compositions) {
+        this.compositions = compositions;
+    }
+
+    public List<ContributionDto> getContributions() {
+        return contributions;
+    }
+
+    public void setContributions(List<ContributionDto> contributions) {
+        this.contributions = contributions;
+    }
 
 }
